Move trends helper
