+++ cellrank/plotting/_gene_trend.py
@@ -13,16 +13,16 @@ import scanpy as sc
 from anndata import AnnData
 from scanpy import logging as logg
 
-from cellrank.plotting._utils import _is_any_gam_mgcv, _fit, _create_models, _model_type
+from cellrank.plotting._utils import (
+    _trends_helper,
+    _is_any_gam_mgcv,
+    _fit,
+    _create_models,
+    _model_type,
+)
 from cellrank.tools._constants import LinKey
 from cellrank.utils._parallelize import parallelize
-from cellrank.utils._utils import (
-    _trends_helper,
-    check_collection,
-    _make_unique,
-    _get_n_cores,
-    save_fig,
-)
+from cellrank.utils._utils import check_collection, _make_unique, _get_n_cores, save_fig
 
 
 def gene_trends(
